filter: factor rules into its own file
